+++ src/config/amqp_qos.rs
@@ -3,12 +3,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize)]
 pub struct AmqpQos {
-    prefetch_count: u32,
+    prefetch_count: u16,
     options: BasicQosOptions,
 }
 
 impl AmqpQos {
-    pub fn prefetch_count(&self) -> u32 {
+    pub fn prefetch_count(&self) -> u16 {
         self.prefetch_count
     }
